[pages/cloud][m]: cloud landing page
