+++ pages/cloud/index.tsx
@@ -1,23 +1,41 @@
 import React from "react";
 import { GetStaticProps, GetStaticPropsResult } from "next";
 
-import { Hero } from "../components/Hero";
-import Features from "../components/Features";
+import { Hero } from "@/components/Hero";
+import { Button } from "@/components/common/Button"
+import { Features } from "@/components/Features";
 
-import type { CustomAppProps } from "./_app";
+import type { CustomAppProps } from "../_app";
+
 
 type Props = CustomAppProps;
 
-const highlights = [
-    "A",
-    "B",
-    "C"
-]
+const heroProps = {
+    title: "Avoid the hassle and complexity of deploying yourself.",
+    subtitle: "With Flowershow Cloud you can turn your markdown into a website in a couple of clicks.",
+    highlights: [
+        "All of Flowershow goodness",
+        "Always up to date",
+        "Run off GitHub",
+    ],
+    infoBadges: [
+        (<h2 className="max-w-lg text-3xl font-bold tracking-tight text-secondary dark:text-white sm:text-4xl">
+            Flowershow Cloud
+        </h2>)
+    ],
+    ctaButtons: [
+        (<Button href="/cloud/waitlist">
+            <span>Get started</span>
+        </Button>),
+    ]
+}
+
+const features = []
 
 export default function Home() {
     return (
         <main>
-            <Hero highlights={highlights} />
+            <Hero {...heroProps} />
 
             <div className="py-10 sm:px-2 lg:relative lg:px-0" id="overview">
                 <div className="prose dark:prose-invert mx-auto max-w-6xl px-4 lg:max-w-6xl lg:px-8 xl:px-12">
@@ -31,7 +49,7 @@ export default function Home() {
                     </p>
                 </div>
             </div>
-            <Features />
+            <Features features={features} />
         </main>
     );
 }
