+++ pages/cloud/waitlist.tsx
@@ -1,0 +1,5 @@
+export default function Page() {
+    return (
+        <h1>Sign up for waitlist</h1>
+    )
+}
